% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.pdc.R
\name{read.pdc}
\alias{read.pdc}
\title{Read Pandas complex numbers into R from a csv file}
\usage{
read.pdc(path, nchars = 1e+07, ...)
}
\arguments{
\item{path}{Path to the csv file}

\item{nchars}{Expected number of characters in csv file, passed to readChar}

\item{...}{Arguments passed to read.csv after the data is cleaned.}
}
\value{
The complex data, having been read into R.
}
\description{
The only function in this package, developed in the aftermath of
realizing that I had written complex numbers to a file that neither
Python (inc. Pandas!) nor R could read by default.
}
\examples{
complexData <- read.pdc("./data/complexData.csv") 
}
